src/main/java/lms/spring61ref/prototype/payment/PaymentService.java
package lms.spring61ref.prototype.payment;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;
import java.math.BigDecimal;
import java.time.LocalDateTime;

@Component
@RequiredArgsConstructor
public class PaymentService  {
    private final ExRateProvider exRateProvider;

    public Payment prepare(Long orderId, String currency, BigDecimal foreignCurrencyAmount)   {
        BigDecimal exRate = exRateProvider.getExRate(currency);
        BigDecimal convertedAmount = foreignCurrencyAmount.multiply(exRate);
        LocalDateTime validUntil = LocalDateTime.now().plusHours(1);
        return new Payment(orderId, currency, foreignCurrencyAmount, exRate, convertedAmount, validUntil);
    }

}
